+++ packages/test-website/src/pages/home/Error.tsx
@@ -1,5 +1,5 @@
 ﻿import { ReactElement, useEffect } from 'react'
-import { cleanUpImproperlyClosedDialog, IError } from '@interface-technologies/iti-react'
+import { IError } from '@interface-technologies/iti-react'
 import { useReady } from 'components/routing'
 import { ErrorType } from '_util/errorHandling'
 
@@ -15,8 +15,6 @@ export default function Page({ error }: ErrorPageProps): ReactElement | null {
             title: 'Error',
             activeNavbarLink: undefined,
         })
-
-        cleanUpImproperlyClosedDialog()
     }, [onReady])
 
     if (!error) return null